Added commenting of errors (#11)
Added a call to the comment function inside the `catch` block.

Just to be sure, we are logging the error first in the case that the
comment fails.

The intention is that, if it fails for a configuration reason, the bot
will be able to report that. (If it fails because of the commenting
feature then it won't be able to report it).

A good example is that it is set for `MERGE` but it is not allowed in
the repository (or `auto-merge` is disabled).

This resolves #2